pyproject.toml
[project]
name = "global-arbitrage"
version = "0.1.0"
description = "Global arbitrage scanner and paper-trading lab focused on Brazil and cross-market dislocations."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Marcelo Nogueira" }]

dependencies = [
    "beautifulsoup4>=4.12",
    "duckdb>=1.1",
    "numpy>=1.26",
    "pandas>=2.2",
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "PyYAML>=6.0",
    "python-dotenv>=1.0",
    "requests>=2.32",
    "rich>=13.7",
    "tenacity>=8.2",
    "yfinance>=0.2.54",
]

[project.optional-dependencies]
mt5 = ["MetaTrader5>=5.0.45"]
dev = [
    "mypy>=1.10",
    "pytest>=8.2",
    "ruff>=0.5",
]

[project.scripts]
ga-scan = "global_arbitrage.cli.scan:main"
ga-backtest = "global_arbitrage.cli.backtest:main"
ga-paper = "global_arbitrage.cli.paper:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/global_arbitrage"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "W"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
